% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_merge.R
\name{de_genes_pairs_multiple}
\alias{de_genes_pairs_multiple}
\title{Title}
\usage{
de_genes_pairs_multiple(
  dat.list,
  de.param.list,
  cl,
  pairs,
  cl.means.list,
  cl.present.list,
  cl.sqr.means.list,
  lfc.conservation.th = 0.6,
  de.genes.list = NULL,
  max.cl.size = 200,
  mc.cores = 1,
  method = "fast_limma"
)
}
\arguments{
\item{dat.list}{}

\item{de.param.list}{}

\item{cl}{}

\item{pairs}{}

\item{cl.means.list}{}

\item{cl.present.list}{}

\item{lfc.conservation.th}{}

\item{de.genes.list}{}

\item{max.cl.size}{}
}
\description{
Title
}
