% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_3d_umap_anno}
\alias{plot_3d_umap_anno}
\title{plot_3d_umap_anno}
\usage{
plots <- plot_3d_umap_anno(umap.fn="3d.umap.sampled.csv", anno.df, cols=c("platform","cluster"),label_cols=list(NULL,NULL),html.fn = "3d.umap",dest.d="data/rd_plot_example" )
}
\arguments{
\item{umap.fn}{path to umap coordinates. CSV file containing sample_id and umap x/y/z coordinates}

\item{anno.df}{Sample annotations. The first column should be sample_id, and each annotation should have \_id, \_label, and \_color columns. Requires cluster_id which needs to be sequential in order of the dendrogram.}

\item{cols}{base name of variables to be represented as bargraphs below dendrogram. Annotation variables need to be represented as \_id, \_label, \_color in anno.}

\item{label_cols}{Which label on top of plot.}

\item{win.dim}{dimensions of plot window}

\item{cex}{size of plotted points. Default = 0.25}

\item{html.fn}{name of html file to save figures can be saved as "png", "pdf" or "both"}

\item{dest.d}{path to load data from and save plots to.}
}
\description{
plot_3d_umap_anno
}
