% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.parquet.R
\name{select_markers_pair_group_top_ds}
\alias{select_markers_pair_group_top_ds}
\title{Title}
\usage{
select_markers_pair_group_top_ds(
  g1,
  g2,
  ds,
  genes,
  cl.bin,
  select.sign = "up",
  n.markers = 20,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{g1}{}

\item{g2}{}

\item{n.markers}{}

\item{cl}{}

\item{top.n}{}

\item{max.num}{}

\item{up.gene.score}{}

\item{down.gene.score}{}
}
\description{
Title
}
